+++ src/Symfony/Component/Messenger/Tests/Command/FailedMessagesShowCommandTest.php
@@ -216,6 +216,54 @@ EOF
         $this->assertStringContainsString('Showing first 1 messages.', $tester->getDisplay(true));
     }
 
+    public function testListMessagesReturnsFilteredByClassMessage()
+    {
+        $sentToFailureStamp = new SentToFailureTransportStamp('async');
+        $envelope = new Envelope(new \stdClass(), [
+            new TransportMessageIdStamp(15),
+            $sentToFailureStamp,
+            new RedeliveryStamp(0),
+            ErrorDetailsStamp::create(new \RuntimeException('Things are bad!')),
+        ]);
+        $receiver = $this->createMock(ListableReceiverInterface::class);
+        $receiver->method('all')->with()->willReturn([$envelope]);
+
+        $command = new FailedMessagesShowCommand(
+            'failure_receiver',
+            $receiver
+        );
+
+        $tester = new CommandTester($command);
+        $tester->execute([]);
+        $this->assertStringContainsString('Things are bad!', $tester->getDisplay(true));
+        $tester->execute(['--class-filter' => 'stdClass']);
+        $this->assertStringContainsString('Things are bad!', $tester->getDisplay(true));
+        $tester->execute(['--class-filter' => 'namespace\otherClass']);
+        $this->assertStringContainsString('[OK] No failed messages were found.', $tester->getDisplay(true));
+    }
+
+    public function testListMessagesReturnsCountByClassName()
+    {
+        $sentToFailureStamp = new SentToFailureTransportStamp('async');
+        $envelope = new Envelope(new \stdClass(), [
+            new TransportMessageIdStamp(15),
+            $sentToFailureStamp,
+            new RedeliveryStamp(0),
+            ErrorDetailsStamp::create(new \RuntimeException('Things are bad!')),
+        ]);
+        $receiver = $this->createMock(ListableReceiverInterface::class);
+        $receiver->method('all')->with()->willReturn([$envelope, $envelope]);
+
+        $command = new FailedMessagesShowCommand(
+            'failure_receiver',
+            $receiver
+        );
+
+        $tester = new CommandTester($command);
+        $tester->execute(['--stats' => 1]);
+        $this->assertStringContainsString('stdClass   2', $tester->getDisplay(true));
+    }
+
     public function testInvalidMessagesThrowsException()
     {
         $sentToFailureStamp = new SentToFailureTransportStamp('async');