+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * `InMemoryTransport` can perform message serialization through dsn `in-memory://?serialize=true`.
+ * Add the options `stats` and `class-filter` to `FailedMessagesShowCommand`
 
 5.2.0
 -----